+++ src/Recyclarr.Core/Config/Models/ServiceConfiguration.cs
@@ -53,6 +53,7 @@ public record CustomFormatGroupConfig
     public string TrashId { get; init; } = "";
     public ICollection<CfGroupAssignScoresToConfig> AssignScoresTo { get; init; } = [];
     public ICollection<string> Select { get; init; } = [];
+    public ICollection<string> Exclude { get; init; } = [];
 }
 
 [UsedImplicitly(ImplicitUseKindFlags.Assign, ImplicitUseTargetFlags.WithMembers)]